% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{buildTree}
\alias{buildTree}
\title{buildTree}
\usage{
buildTree(object, method = "euclidean", dim.type = "umap",
  dim.use = 1:2, verbose = FALSE)
}
\arguments{
\item{object}{an FSPY object}

\item{method}{character. Mehtod to build MST.}

\item{dim.type}{character.}

\item{dim.use}{numeric}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
buildTree
}
\examples{

if (FALSE) {
# build minimum spanning tree (MST) based on tsne
fspy <- buildTree(fspy, dim.type = "tsne", dim.use = 1:2)

# Using PCA
fspy <- buildTree(fspy, dim.type = "pca", dim.use = 1:4)

# Using UMAP
fspy <- buildTree(fspy, dim.type = "umap", dim.use = 1:2)

# Using Diffusion Maps
fspy <- buildTree(fspy, dim.type = "dc", dim.use = 1:3)
}

}
