% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constraintMatrix}
\alias{constraintMatrix}
\title{constraintMatrix}
\usage{
constraintMatrix(x, cutoff = 0.99, markers = NULL,
  method = "euclidean")
}
\arguments{
\item{x}{matrix}

\item{cutoff}{numeric. Cutoff of the constraint value}

\item{markers}{character. Markers used in the calculation of constraint model.}

\item{method}{character. the distance measure to be used.
This must be one of "euclidean", "maximum", "manhattan",
"canberra", "binary" or "minkowski".}
}
\description{
constraint FCS data by a provid cutoff
}
