% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{correctBatchFSPY}
\alias{correctBatchFSPY}
\title{correctBatchFSPY}
\usage{
correctBatchFSPY(object, batch = NULL, verbose = F)
}
\arguments{
\item{object}{An FSPY object}

\item{batch}{vector.}

\item{...}{Parameters passing to \code{\link[ComBat]{sva}} function}
}
\value{
An FSPY object after removing batch effect
}
\description{
Remove batch effect in FSPY object
}
\seealso{
\code{\link[BiocNeighbors]{findKNN}}
}
