% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{createFSPY}
\alias{createFSPY}
\title{create an FSPY object}
\usage{
createFSPY(raw.data, markers, meta.data, batch = NULL,
  batch.correct = FALSE, normalization.method = c("log", "none"),
  verbose = FALSE, ...)
}
\arguments{
\item{raw.data}{matrix. Raw data read from FCS file after perform preprocessing.}

\item{markers}{vector. Detailed marker information in the gate of flow cytometer.}

\item{meta.data}{data.frame. Raw metadata of each cell. Columns "cell" and "stage" are required.}

\item{normalization.method}{character. Normalization and transformation method.
Whether to log transformed raw.data. If FALSE, it's better
to perform transformation using \code{\link[transformation]{flowCore}} before creating FSPY
object. flowSpy only provide log transforma parameter. If you need to using truncateTransform,
scaleTransform, linearTransform, quadraticTransform and lnTransform, see \code{flowCore} for more
information.}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{paramters pass to \code{correctBatchFSPY} function.}
}
\value{
An FSPY object with raw.data and markers and meta.data
}
\description{
This function include all information used in flow cytometry data
processing.
}
\examples{
# See vignette tutorial
vignette(package = "flowSpy")
vignette("Quick_start", package = "flowSpy")
vignette("Tutorial_of_flowSpy", package = "flowSpy")



}
