% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defLeafCells}
\alias{defLeafCells}
\title{definition of leaf cells}
\usage{
defLeafCells(object, leaf.cells = NULL, pseudotime.cutoff = 0,
  verbose = T)
}
\arguments{
\item{object}{an FSPY object}

\item{leaf.cells}{character or numeric. Cell name of the root cells or cluster.id of root.cells}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
definition of root cells
}
