% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defRootCells}
\alias{defRootCells}
\title{definition of root cells}
\usage{
defRootCells(object, root.cells = NULL, verbose = F)
}
\arguments{
\item{object}{an FSPY object}

\item{root.cells}{vector. Cell name of the root cells}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
definition of root cells
}
