% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchCell}
\alias{fetchCell}
\title{Fetching cellls of FSPY}
\usage{
fetchCell(object, logical.connect = "or", verbose = F, ...)
}
\arguments{
\item{object}{An FSPY object}

\item{logical.connect}{character. "and" or "or"}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Paramters to pass to limitation}
}
\description{
Fetching cellls of FSPY
}
