% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchClustMeta}
\alias{fetchClustMeta}
\title{Fetching clusters' metadata of FSPY}
\usage{
fetchClustMeta(object, verbose = F)
}
\arguments{
\item{object}{An FSPY object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\description{
Fetching clusters' metadata of FSPY
}
