% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FSPY}
\alias{FSPY}
\alias{FSPYclass}
\title{flowSpy class}
\description{
All information stored in FSPY object, and you can use creatFSPY to
create this object. In this package, most of the function will use
FSPY object as input, and return a modified SPTY obejct as well.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{matrix, raw signal data captured using flow cytometry.}

\item{\code{raw.meta.data}}{data.frame, raw meta data information from the experiment.}

\item{\code{log.data}}{matrix, data set log transfromed from the raw.data.}

\item{\code{gate.data}}{matrix, data set used in the calculation of PCA, tSNE and destiny.}

\item{\code{meta.data}}{data.frame, meta data information after gating.}

\item{\code{markers}}{vector, markers used in the calculation of PCA, tSNE and destiny.}

\item{\code{markers.idx}}{vector, index of markers used in the calculation of PCA, tSNE and destiny.}

\item{\code{gating}}{data.frame, gating parameter stored in the FSPY object.}

\item{\code{pca.sdev,pca.load,pca.score}}{pca information of FSPY object which generated from \code{\link[gmodels]{fast.prcomp}}.}

\item{\code{tsne.value}}{data.frame, tSNE coordinates information of FSPY object.}

\item{\code{dm}}{data.frame, destiny information of FSPY object.}

\item{\code{som}}{list, som network of FSPY object.}

\item{\code{cluster}}{data.frame, cluster in FSPY.}

\item{\code{root.cell}}{vector, names of root cells.}

\item{\code{leaf.cell}}{vector, names of leaf cells.}

\item{\code{pseudotime}}{data.frame, pseudotime of all cells.}

\item{\code{walk}}{list, random forward and backward walk between tips and cells.}

\item{\code{diff.tree}}{list, differentiation tree of FSPY object.}

\item{\code{diff.traj}}{list, differentiation trajectory of FSPY object.}

\item{\code{network}}{list, network of FSPY object.}

\item{\code{sankey}}{list, sankey diagrams of FSPY object.}
}}

