% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FSPYclass}
\alias{FSPYclass}
\alias{FSPY}
\title{flowSpy class}
\description{
All information stored in FSPY object, and you can use creatFSPY to
create this object. In this package, most of the function will use
FSPY object as input, and return a modified SPTY obejct as well.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{matrix. Raw signal data captured using flow cytometry.}

\item{\code{log.data}}{matrix. Log-transfromed dataset of raw.data.}

\item{\code{meta.data}}{data.frame. Meta data information.}

\item{\code{markers}}{vector. Markers used in the calculation of PCA, tSNE, destiny and umap.}

\item{\code{markers.idx}}{vector. Index of markers used in the calculation of PCA, tSNE, destiny and umap.}

\item{\code{cell.name}}{vector. Cell names of log data.}

\item{\code{pca.sdev,pca.load,pca.scores}}{Pca information of FSPY object which are generated from \code{\link[gmodels]{fast.prcomp}}.}

\item{\code{knn}}{numeric. Numbers of nearest neighbors}

\item{\code{knn.index,knn.distance}}{matrix. Each row of the \code{knn.index} matrix corresponds to a point
in \code{log.data} and contains the row indices in \code{log.data} that are its nearest neighbors.
And each row of the \code{knn.distance} contains the distance of its nearest neighbors.}

\item{\code{tsne.parameter}}{list. tSNE paramters in calculation of the \code{tsne.value} in \code{\link[Rtsne]{Rtsne}}}

\item{\code{tsne.value}}{matrix. tSNE coordinates information.}

\item{\code{dm.parameter}}{list. Diffusion map parameters in calculation of the \code{dm}.}

\item{\code{dm}}{data.frame. Diffusion map calculated by \code{\link[destiny]{DiffusionMap}}}

\item{\code{som}}{list. Som network calculate by \code{\link[FlowSOM]{FlowSOM}}.}

\item{\code{umap.layout}}{matrix umap coordinates information calculated using \code{\link[umap]{umap}}.}

\item{\code{root.cells}}{vector, Names of root cells.}

\item{\code{leaf.cells}}{vector. Names of leaf cells.}

\item{\code{pseudotime}}{data.frame. Pseudotime of all cells.}

\item{\code{walk}}{list. Random forward and backward walk between \code{root.cells} and \code{leaf.cells}.}

\item{\code{diff.tree}}{list. Differentiation tree of all cells.}

\item{\code{diff.traj}}{list. Differentiation trajectory all cells.}

\item{\code{network}}{list. Network.}

\item{\code{plot.meta}}{data.frame. Plot meta information for \code{plot2D} or \code{plot3D}.}

\item{\code{add.meta}}{list. Additional meta information of FSPY object.}
}}

