% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSPYdata.R
\docType{data}
\name{FSPYdata}
\alias{FSPYdata}
\alias{data:FSPYdata}
\title{Time course flow cytometry data, a matrix containing expression matrix,
and a data.frame containing meta data to build an FSPY object}
\format{An FSPY object.}
\usage{
data(FSPYdata)
}
\value{
an FSPY object, .
}
\description{
Time course flow cytometry data, a matrix containing expression matrix,
and a data.frame containing meta data to build an FSPY object
}
\author{
Yuting Dai \email{forlynna@sjtu.edu.cn}
}
\keyword{data}
