% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garbage.R
\name{garbage}
\alias{garbage}
\title{A single iteration of flood pseudotime calculation.}
\usage{
garbage(object)
}
\arguments{
\item{object}{An FSPY object}

\item{minimum.cells.flooded}{(Numeric) Stop pseudotime calculation when fewer cells are newly visited in a given iteration, and assign remaining unvisited cells pseudotime NA.}

\item{tm.flood}{(Sparse or full matrix) A sparse matrix of normalized transition probabilities. If unprovided (\code{NULL}), this will be calculated automatically.}

\item{x}{(Numeric Vector) Probabilities}
}
\value{
(Numeric vector) The iteration that newly visited each cell.

(Numeric) Cumulative probability
}
\description{
A single iteration of flood pseudotime calculation.
}
\references{
This code is strongly based on the \code{\link[URD]{floodPseudotime}} function,
  which is developed by Jeffrey A. Farrell. See paper Jeffrey A. Farrell, et. al.,
  Single-cell reconstruction of developmental trajectories during zebrafish
  embryogenesis, Science, 2018.


Combine probabilities

Calculates cumulative probability
}
\keyword{internal}
