% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gatingFSPY}
\alias{gatingFSPY}
\title{Apply gating on FSPY object}
\usage{
gatingFSPY(object, gating = NULL, initialization = T, verbose = T)
}
\arguments{
\item{object}{S4 object. FSPY object}

\item{gating}{data.frame. gating paramete, the first column in gating is the marker name,
the second column is the start paramter of gating cutoff, and the third column
is the end paramter of gating cutoff.}

\item{initialization}{logical. Whether to initialize gating process}

\item{verbose}{logical. Whether to print calculation progress.}
}
\description{
Apply gating on FSPY object
}
