% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plot2D}
\alias{plot2D}
\title{Visualization of 2D data}
\usage{
plot2D(object, item.use = c("PC1", "PC2"), color.by = "stage",
  size = 1, alpha = 1, main = "2D plot of FSPY",
  color.theme = c("#f2de00", "#8bd129", "#33CCFF", "#0066CC", "#CC66FF",
  "#FF3300"))
}
\arguments{
\item{object}{An FSPY object}

\item{item.use}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of raw.meta.data, or it can be just "density" (the default value).}

\item{size}{numeric. size of the dot}

\item{alpha}{numberic. transparency (0-1) of the dot, default is 0.2.}

\item{main}{character. title of the plot}

\item{color.theme}{character. Library of color theme}
}
\description{
Visualization of 2D data
}
\examples{

}
