% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{Visualization of 3D data of FSPY}
\usage{
plot3D(object, item.use = c("PC1", "PC2", "PC3"), color.by = "stage",
  order.by = NULL, size = 1, angle = 60, scale.y = 0.8,
  category = "categorical", main = "3D plot of FSPY", ...)
}
\arguments{
\item{object}{An FSPY object}

\item{item.use}{character. Items use to 3D plot, axes x and y and z must be numeric.}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}

\item{order.by}{character. Order of color theme.}

\item{size}{numeric. size of the dot}

\item{angle}{numberic. angle of the plot}

\item{scale.y}{numeric. scale of y axis related to x- and z axis}

\item{category}{character. numeric or categorical}

\item{main}{character. title of the plot}

\item{...}{options to pass on to the \code{\link[scatterplot3d]{scatterplot3d}} function.}
}
\description{
Visualization of 3D data of FSPY
}
