% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Visualization heatmap of cluster data of FSPY}
\usage{
plotClusterHeatmap(object, color = colorRampPalette(c("blue", "white",
  "red"))(100), scale = "row", ...)
}
\arguments{
\item{object}{An FSPY object}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}
}
\description{
Visualization heatmap of cluster data of FSPY
}
\examples{

plotClusterHeatmap(fspy)
plotClusterHeatmap(fspy, color = colorRampPalette(c("purple","white","yellow"))(100))
plotClusterHeatmap(fspy, cluster_row = F)
plotClusterHeatmap(fspy, cluster_row = F, cluster_col = F)

}
