% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plotGATE}
\alias{plotGATE}
\title{plot with gate}
\usage{
plotGATE(object, plot.markers, plot.type = c("dot", "mesh", "dotmesh",
  "dotellipse", "ellipse"), color.by = "density", size = 1,
  alpha = 0.2, color.theme = c("#666666", "#f2de00", "#8bd129",
  "#33CCFF", "#0066CC", "#CC66FF", "#FF3300"), gated = T,
  show.gate = F)
}
\arguments{
\item{object}{S4 object. FSPY object}

\item{plot.markers}{vector included 2 charaters. For example c("CD34", "CD43").}

\item{plot.type}{character. Plot style for flow cytometry data. default: dot.}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of raw.meta.data, or it can be just "density" (the default value).}

\item{size}{numeric. size of the dot}

\item{alpha}{numberic. transparency (0-1) of the dot, default is 0.2.}

\item{color.theme}{character. Library of color theme}

\item{gated}{logical. Whether to show only gated data.}

\item{show.gate}{logical. Whether to show gated cells. This parameter will only take function when
\code{color.by} is "density"}
}
\description{
plot with gate
}
\examples{

}
