% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Visualization heatmap of data of FSPY}
\usage{
plotHeatmap(object, color = colorRampPalette(c("blue", "white",
  "red"))(100), scale = "row", downsize = 1000, cluster_rows = F,
  cluster_cols = F, ...)
}
\arguments{
\item{object}{An FSPY object}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}
}
\description{
Visualization heatmap of data of FSPY
}
\examples{

}
