% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKNNtree.R
\name{plotKNNtree}
\alias{plotKNNtree}
\title{plot Minimum spanning tree from knn cluster of FSPY}
\usage{
plotKNNtree(object, cex.size = 1, color.by = "default",
  size.by = "cell.percent", layout = "mst", show.node.name = F,
  color.theme = NULL)
}
\arguments{
\item{object}{an FSPY object}

\item{cex.size}{numeric. size cex of the dot}

\item{color.by}{numeric. size color theme of the dot}

\item{size.by}{numeric. size theme of the dot}

\item{show.node.name}{logical. whether to show node name}

\item{color.theme}{character. Library of color theme}
}
\description{
plot Minimum spanning tree from knn cluster of FSPY
}
