% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plotMarkerDensity}
\alias{plotMarkerDensity}
\title{plotMarkerDensity}
\usage{
plotMarkerDensity(object, cutoff = -1, markers = NULL, adjust = 0.5,
  var.cols = T, plot.theme = theme_bw())
}
\arguments{
\item{object}{An FSPY object}

\item{cutoff}{numeric. Cutoff of trajectory value}

\item{markers}{character. Markers used in the calculation progress}

\item{adjust}{numeric. Transparency (0-1) of the dot, default is 1.}

\item{var.cols}{logical. Whether to plot stage}

\item{plot.theme}{themes from \code{ggplot2}}
}
\description{
plotMarkerDensity
}
