% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotPieTree}
\alias{plotPieTree}
\title{plot MST pie of FSPY}
\usage{
plotPieTree(object, cex.size = 2, size.by.cell.number = T,
  as.tree = F, root.id = NULL, show.node.name = F,
  color.theme = NULL)
}
\arguments{
\item{object}{an FSPY object}

\item{cex.size}{numeric. size cex of the dot}

\item{size.by.cell.number}{logical.}

\item{show.node.name}{logical. whether to show node name}

\item{color.theme}{character. Library of color theme}
}
\description{
plot MST pie of FSPY
}
\examples{
# Runs only have two or more stages
plotPieTree(fspy, cex.size = 1, size.by.cell.number = T) +
   scale_fill_manual(values = c("#00599F","#FF3222","#009900",
                                "#FF9933","#FF99FF","#7A06A0"))

}
