% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKNNtree.R, R/plotSOM.R
\name{plotSOMtree}
\alias{plotSOMtree}
\title{plot Minimum spanning tree from knn cluster of FSPY}
\usage{
plotSOMtree(object, root.id = NULL, cex.size = 1,
  color.by = "default", size.by = "cell.percent", show.node.name = F,
  color.theme = NULL)

plotSOMtree(object, root.id = NULL, cex.size = 1,
  color.by = "default", size.by = "cell.percent", show.node.name = F,
  color.theme = NULL)
}
\arguments{
\item{object}{an FSPY object}

\item{cex.size}{numeric. size cex of the dot}

\item{color.by}{numeric. size color theme of the dot}

\item{size.by}{numeric. size theme of the dot}

\item{show.node.name}{logical. whether to show node name}

\item{color.theme}{character. Library of color theme}

\item{object}{an FSPY object}

\item{color.by}{character.}

\item{main}{character. title of the plot}

\item{color.theme}{character. Library of color theme}
}
\description{
plot Minimum spanning tree from knn cluster of FSPY

plot som tree of FSPY
}
\examples{



}
