% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTree.R
\name{plotTree}
\alias{plotTree}
\title{plot MST of FSPY}
\usage{
plotTree(object, cex.size = 1, color.by = "cell.number",
  size.by = "cell.number", as.tree = F, root.id = NULL,
  show.node.name = F, color.theme = NULL)
}
\arguments{
\item{object}{an FSPY object}

\item{cex.size}{numeric. size cex of the dot}

\item{color.by}{numeric. size color theme of the dot}

\item{size.by}{numeric. size theme of the dot}

\item{show.node.name}{logical. whether to show node name}

\item{color.theme}{character. Library of color theme}
}
\description{
plot MST of FSPY
}
