% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotViolin}
\alias{plotViolin}
\title{Visualization violin plot of FSPY}
\usage{
plotViolin(object, marker, color.by = "cluster.id", order.by = NULL,
  size = 1, text.angle = 0, main = "Violin plot FSPY",
  plot.theme = theme_bw())
}
\arguments{
\item{object}{An FSPY object}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}

\item{order.by}{vector. Order of color theme.}

\item{size}{numeric. Size of the dot}

\item{main}{character. Title of the plot.}

\item{plot.theme}{themes from \code{ggplot2}}

\item{item.use}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{alpha}{numberic. Transparency (0-1) of the dot, default is 1.}

\item{category}{character. numeric or categorical}

\item{show.cluser.id}{logical. Whether to show cluster id in the plot.}

\item{show.cluser.id.size}{numeric. Size of the cluster id.}
}
\description{
Visualization violin plot of FSPY
}
\examples{

plotViolin(fspy, marker = "CD34")

}
