% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{processingCluster}
\alias{processingCluster}
\title{processingCluster}
\usage{
processingCluster(object, perplexity = 5, k = 5,
  umap.config = umap.defaults, verbose = F, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the clustering functions.}
}
\value{
An FSPY object with cluster.id in meta.data
}
\description{
Compute a specific clustering using the combined flow
   cytometry data. "som" \code{\link[flowSOM]{SOM}}, "hclust" \code{\link[stats]{hclust}},
   "mclust" \code{\link[mclust]{mlcust}}, "kmeans" \code{\link[stats]{kmeans}} are
   provided.
}
\seealso{
Four clustering methods are provided: \code{\link[flowSOM]{SOM}}, \code{\link[stats]{hclust}},
   \code{\link[mclust]{mlcust}}, \code{\link[stats]{kmeans}}. You can use \code{runSOM}, \code{runHclust},
   \code{runMclust} and \code{runKmeans} to run clustering respectively.
}
