% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{pseudotimeFlood}
\alias{pseudotimeFlood}
\title{Calculate pseudotime by 'flooding'}
\usage{
pseudotimeFlood(object, n = 10, minimum.cells.flooded = 2,
  thread = 1, verbose = F)
}
\description{
This calculates pseudotime by performing a probabilistic breadth-first search
of the k-nearest neighbor graph that was used to generate the diffusion map
calculated on the data. The results of this function should be passed to
\code{\link{floodPseudotimeProcess}} to convert them into pseudotime.
}
\references{
This code is strongly based on the \code{\link[URD]{floodPseudotime}} function,
  which is developed by Jeffrey A. Farrell. See paper Jeffrey A. Farrell, et. al.,
  Single-cell reconstruction of developmental trajectories during zebrafish
  embryogenesis, Science, 2018.
}
\keyword{internal}
