% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{pseudotimeProcess}
\alias{pseudotimeProcess}
\title{Process Flood Pseudotime}
\usage{
pseudotimeProcess(object, pseudotime.fun = mean, stability.div = 10,
  n = 10, minimum.cells.flooded = 2, thread = 1, verbose = F)
}
\description{
This processes the values returned by \code{\link{floodPseudotime}} and stores
the result as a pseudotime in an URD object.
}
\examples{

}
\references{
This code is strongly based on the \code{\link[URD]{floodPseudotime}} function,
  which is developed by Jeffrey A. Farrell. See paper Jeffrey A. Farrell, et. al.,
  Single-cell reconstruction of developmental trajectories during zebrafish
  embryogenesis, Science, 2018.
}
