% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runCluster}
\alias{runCluster}
\title{Specific Clustering Method Toolkits}
\usage{
runCluster(object, cluster.method = "som", verbose = T, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{cluster.method}{character. Four clustering method are provided: som, hclust, mclust, kmeans.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the clustering functions.}
}
\value{
An FSPY object with cluster.id in meta.data
}
\description{
Compute a specific clustering using the combined flow
   cytometry data. "som" \code{\link[flowSOM]{SOM}}, "hclust" \code{\link[stats]{hclust}},
   "mclust" \code{\link[mclust]{mlcust}}, "kmeans" \code{\link[stats]{kmeans}} are
   provided.
}
\seealso{
Four clustering methods are provided: \code{\link[flowSOM]{SOM}}, \code{\link[stats]{hclust}},
   \code{\link[mclust]{mlcust}}, \code{\link[stats]{kmeans}}. You can use \code{runSOM}, \code{runHclust},
   \code{runMclust} and \code{runKmeans} to run clustering respectively.
}
