% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{runCompensation}
\alias{runCompensation}
\title{runCompensation}
\usage{
runCompensation(x, spillover = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link[flowCore]{flowFrame}} or \code{\link[flowCore]{flowSet}}.}

\item{spillover}{matrix, The spillover or compensation matrix.}

\item{...}{options to pass on to the further arguments. See \code{\link[flowCore]{compensate}}}
}
\description{
Run compensation by applying a spillover matrix.
   See \code{\link[flowCore]{compensate}} for more information.
}
\examples{

# DO NOT RUN
if (F) {
  fileName <- system.file("extdata","D2.fcs",
                          package="flowSpy")

  # read FACs data
  fcs.data <- read.FCS(filename = fileName)

  # run compensation
  fcs.data <- runCompensation(fcs.data)

}

}
\seealso{
\code{\link[flowCore]{compensate}}
}
