% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destiny.R
\name{runDiffusionMap}
\alias{runDiffusionMap}
\title{Calculate diffusion map in FSPY}
\usage{
runDiffusionMap(object, sigma.use = NULL, distance = c("euclidean",
  "cosine", "rankcor"), density.norm = TRUE, verbose = T, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{sigma.use}{numeric. Diffusion scale parameter of the Gaussian kernel. One of '\code{local}',
 '\code{global}', a \code{\link[base]{numeric}} global sigma or a Sigmas object.
 When choosing '\code{global}', a global sigma will be calculated using find_sigmas
 (See \code{\link[destiny]{find_sigmas}}). A larger sigma might be necessary if the eigenvalues can not
be found because of a singularity in the matrix. See \code{\link[destiny]{DiffusionMap}}}

\item{distance}{Distance measurement method applied to data or a distance matrix/dist.
For the allowed values, see \code{\link[destiny]{find_knn}}}

\item{density.norm}{logical. If TRUE, use density normalisation. See \code{\link[destiny]{DiffusionMap}}}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{options to pass on to the \code{\link[destiny]{DiffusionMap}} function}
}
\description{
Calculate diffusion map in FSPY
}
\seealso{
\code{\link[destiny]{DiffusionMap}}
}
