% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runKmeans}
\alias{runKmeans}
\title{runKmeans}
\usage{
runKmeans(object, k = 25, iter.max = 10, nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE, scale = T, verbose = T)
}
\arguments{
\item{object}{an FSPY object}
}
\value{
an FSPY object
}
\description{
Perform k-means clustering on a data matrix.
}
\examples{



}
