% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runMclust}
\alias{runMclust}
\title{runMclust}
\usage{
runMclust(object)
}
\arguments{
\item{object}{an FSPY object}
}
\value{
an FSPY object
}
\description{
Model-based clustering based on parameterized finite Gaussian mixture models.
}
