% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runPhenograph}
\alias{runPhenograph}
\title{RphenoGraph clustering}
\usage{
runPhenograph(object, mode = "undirected", knn = 30, verbose = F,
  ...)
}
\arguments{
\item{object}{Input data matrix.}

\item{mode}{Number of nearest neighbours, default is 30.}
}
\value{
a communities object, the operations of this class contains:
}
\description{
R implementation of the phenograph algorithm
   A simple R implementation of the phenograph
   [PhenoGraph](http://www.cell.com/cell/abstract/S0092-8674(15)00637-6) algorithm,
   which is a clustering method designed for high-dimensional single-cell
   data analysis. It works by creating a graph ("network") representing
   phenotypic similarities between cells by calculating the Jaccard
   coefficient between nearest-neighbor sets, and then identifying communities
   using the well known [Louvain method](https://sites.google.com/site/findcommunities/)
   in this graph.
}
