% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{runWalk}
\alias{runWalk}
\title{Walk between root cells and leaf cells}
\usage{
runWalk(object, mode = "undirected", max.run = 20, weighted = T,
  verbose = T, ...)
}
\arguments{
\item{object}{An FSPY object}

\item{mode}{character. Specifies how igraph should interpret the supplied matrix.
Possible values are: directed, undirected, upper, lower, max, min, plus.}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to calculation function.}
}
\value{
An FSPY object
}
\description{
Walk between root cells and leaf cells
}
