% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetFSPY.R
\name{subsetFSPY}
\alias{subsetFSPY}
\title{subset FSPY object}
\usage{
subsetFSPY(object, cells = NULL, id.keep = NULL, id.name = NULL,
  recalculate = F, knn = NA, verbose = F)
}
\arguments{
\item{object}{An FSPY object}

\item{cells}{vector, Names of the cells to retain.}

\item{id.keep}{numeric or character. Name of the id to retain}

\item{id.name}{character. Name of the column which id.keep is in it.}

\item{recalculate}{logic. Whether to recalculate the PCA, tSNE, destiny and umap score.}

\item{knn}{numeric. When recalculate is TRUE.}

\item{verbose}{logic. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
This subsets an FSPY object by given a list of cells or cluster id.
    This function will subset all results without recalculating them, such as knn,
    PCA, tSNE, umap and pseudotime. However, this function will worked after running
    \code{\link[flowSpy]{runKNN}}. For instance, you can choose recalculate PCA and
    tSNE and destiny scores by paramter recalculate.
}
\examples{

}
