% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updatePlotMeta}
\alias{updatePlotMeta}
\title{Update plot metadata of FSPY}
\usage{
updatePlotMeta(object, verbose = TRUE)
}
\arguments{
\item{object}{An FSPY object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\description{
Update plot metadata of FSPY
}
