% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{createFSPY}
\alias{createFSPY}
\title{create an FSPY object}
\usage{
createFSPY(raw.data, markers, meta.data, log.transformed = T,
  verbose = T)
}
\arguments{
\item{raw.data}{matrix. Raw data read from FCS file.}

\item{markers}{vector. Detailed marker information in the gate of flow cytometer}

\item{meta.data}{data.frame. Raw metadata of each cell. column "cell" and "stage" are required}

\item{log.transformed}{logical. Whether to log transformed the raw data. If not, it's better
to set the print parameter as logical.}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
This function include all information used in flow cytometry data
processing.
}
\examples{


}
