% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{defRootCells}
\alias{defRootCells}
\title{definition of root cells}
\usage{
defRootCells(object, root.cells = NULL)
}
\arguments{
\item{object}{an FSPY object}

\item{root.cells}{vector, cell name of the root cells}
}
\description{
definition of root cells
}
\examples{



}
