% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchPlotMeta}
\alias{fetchPlotMeta}
\title{Fetching plot metadata of FSPY}
\usage{
fetchPlotMeta(object, verbose = T)
}
\arguments{
\item{object}{An FSPY object}

\item{verbose}{logical. Whether to print calculation progress.}
}
\description{
Fetching plot metadata of FSPY
}
