% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSpy-package.R
\docType{package}
\name{flowSpy-package}
\alias{flowSpy-package}
\alias{flowSpy}
\title{Visualization and analyzation for flow cytometry data}
\description{
Functions and methods to visualize and analyze flow cytometry data.
}
\details{
\tabular{ll}{ Package: \tab flowSpy\cr Type: \tab Package\cr Version: \tab
0.0.1\cr Date: \tab 2018-09-26\cr License: \tab MIT\cr }
Large parts of the analysis process of flow cytometry data has been mostly
qualitative. To this end, dedicated visualization techniques have been used
for both quality control and infernece of the data. This package provides a
number of different visualization tools for flow data.
}
\examples{

## examples go here


}
\author{
Maintainer: Yuting Dai <forlynna@sjtu.edu.cn>
Authors: Yuting Dai
}
\keyword{package}
