% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2D}
\alias{plot2D}
\title{Visualization of 2D data of FSPY}
\usage{
plot2D(object, item.use = c("PC1", "PC2"), color.by = "stage",
  order.by = NULL, size = 1, alpha = 1, category = NULL,
  main = "2D plot of FSPY", plot.theme = theme_base(),
  color.theme = NULL)
}
\arguments{
\item{object}{An FSPY object}

\item{item.use}{character. Items use to 2D plot, axes x and y must be numeric.}

\item{color.by}{character. Dot or mesh color by which character. It can be one of the column
of plot.meta, or it can be just "density" (the default value).}

\item{size}{numeric. size of the dot}

\item{alpha}{numberic. transparency (0-1) of the dot, default is 0.2.}

\item{category}{character. numeric or categorical}

\item{main}{character. title of the plot}

\item{plot.theme}{themes from \code{\link[ggthemes]{ggthemes}}}

\item{color.theme}{character. Library of color theme}
}
\description{
Visualization of 2D data of FSPY
}
\examples{

}
