% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plot3D}
\alias{plot3D}
\title{Visualization of 2D data}
\usage{
plot3D(object, item.use = c("PC1", "PC2", "PC3"), color.by = "stage",
  scale.y = 0.9, angle = 40, size = 1, main = "3D plot of FSPY",
  color.theme = c("#f2de00", "#8bd129", "#33CCFF", "#0066CC", "#CC66FF",
  "#FF3300"))
}
\arguments{
\item{object}{An FSPY object}

\item{item.use}{character. Items use to 3D plot, axes x and y and z must be numeric.}

\item{color.by}{character. Dot color by which character. It can be one of the column
of raw.meta.data, or it can be just "density" (the default value).}

\item{scale.y}{numeric. scale of y axis related to x- and z axis}

\item{angle}{numberic. angle between x and y axis (Attention: result depends on scaling).}

\item{size}{numeric. size of the dot}

\item{main}{character. title of the plot}

\item{color.theme}{character. Library of color theme}
}
\description{
Visualization of 2D data
}
\examples{

}
