% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Visualization heatmap of data of FSPY}
\usage{
plotHeatmap(object, color = colorRampPalette(c("blue", "white",
  "red"))(100), scale = "row", downsize = 1000, cluster_rows = F,
  cluster_cols = F, ...)
}
\arguments{
\item{object}{An FSPY object}

\item{color}{vector. Colors used in heatmap.}

\item{scale}{character. Whether the values should be centered and scaled in either
the row direction or the column direction, or none. Corresponding values are
"row", "column" and "none"}

\item{downsize}{numeric. Cells size used to plot heatmap}

\item{cluster_rows}{logical. Whether rows should be clustered}

\item{cluster_cols}{logical. Whether columns should be clustered}

\item{...}{options to pass on to the \code{\link[pheatmap]{pheatmap}} function.}
}
\description{
Visualization heatmap of data of FSPY
}
\examples{

if (F) {

plotHeatmap(fspy)
plotHeatmap(fspy, cluster_rows = T)
plotHeatmap(fspy, cluster_rows = T, clustering_method = "ward.D")
plotHeatmap(fspy, cluster_rows = T, cluster_cols = T)

}


}
