% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plotPseudotimeDensity}
\alias{plotPseudotimeDensity}
\title{plot heatmap of FSPY}
\usage{
plotPseudotimeDensity(object, color.by = "stage",
  main = "Density of pseudotime", adjust = 0.5,
  color.theme = c("#f2de00", "#8bd129", "#33CCFF", "#0066CC", "#CC66FF",
  "#FF3300"))
}
\arguments{
\item{object}{an FSPY object}

\item{color.by}{character.}

\item{main}{character. title of the plot}

\item{color.theme}{character. Library of color theme}
}
\description{
plot heatmap of FSPY
}
\examples{

}
