% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFSPY.R
\name{plotPseudotimeDensity}
\alias{plotPseudotimeDensity}
\title{plot Pseudotime density of FSPY}
\usage{
plotPseudotimeDensity(object, color.by = "stage",
  main = "Density of pseudotime", adjust = 0.5,
  plot.theme = theme_bw())
}
\arguments{
\item{object}{an FSPY object}

\item{color.by}{character.}

\item{main}{character. Title of the plot}

\item{adjust}{numeric. A multiplicate bandwidth adjustment.}

\item{plot.theme}{themes from \code{ggplot2}}
}
\description{
plot Pseudotime density of FSPY
}
