% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSOM.R
\name{plotSOMtree}
\alias{plotSOMtree}
\title{plot som tree of FSPY}
\usage{
plotSOMtree(object, root.id = NULL, cex.size = 1,
  color.by = "default", size.by = "cell.percent", show.node.name = F,
  color.theme = NULL)
}
\arguments{
\item{object}{an FSPY object}

\item{color.by}{character.}

\item{color.theme}{character. Library of color theme}

\item{main}{character. title of the plot}
}
\description{
plot som tree of FSPY
}
\examples{



}
