% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runCluster}
\alias{runCluster}
\title{Specific Clustering Method Toolkits}
\usage{
runCluster(object, cluster.method = "som", verbose = T, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{cluster.method}{character. Four clustering method are provided: som, hclust, mclust, kmeans.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the clustering functions.}
}
\value{
An FSPY object
}
\description{
Compute a specific clustering using the combined flow
   cytometry data. "som" \code{\link[SOM]{flowSOM}}, "hclust" \code{\link[hclust]{stats}},
   "mclust" \code{\link[mclust]{mlcust}}, "kmeans" \code{\link[kmeans]{stats}} are
   provided.
}
\examples{




}
