% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runCluster}
\alias{runCluster}
\title{Specific Clustering Method Toolkits}
\usage{
runCluster(object, cluster.method = c("som", "kmeans", "clara",
  "phenograph", "hclust", "mclust"), verbose = F, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{cluster.method}{character. Four clustering method are provided: som, clara, kmeans and phenograph.
Clustering method "hclust" and "mclust" are not recommended because of long computing time.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{...}{options to pass on to the clustering functions.}
}
\value{
An FSPY object with cluster.id in meta.data
}
\description{
Compute a specific clustering using the combined flow
   cytometry data. "som" \code{\link[flowSOM]{SOM}}, "hclust" \code{\link[stats]{hclust}},
   "clara" \code{\link[cluster]{clara}}, "phenograph", "kmeans" \code{\link[stats]{kmeans}} are
   provided.
}
\seealso{
\code{\link[flowSOM]{SOM}}, \code{\link[stats]{hclust}},
   \code{\link[cluster]{clara}}, \code{\link[stats]{kmeans}}.
   You can use \code{runSOM}, \code{runClara},
   \code{runPhenotype}, \code{runKmeans}, \code{runMclust} and
   \code{runHclust} to run clustering respectively.
}
