% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{runKNN}
\alias{runKNN}
\title{Calculate k-nearest neighbors of FSPY}
\usage{
runKNN(object, knn = 30, BNPARAM = NULL, knn.replace = T,
  verbose = F, ...)
}
\arguments{
\item{object}{an FSPY object}

\item{knn}{numeric. Number of k-nearest neighbors.}

\item{BNPARAM}{A BiocNeighborParam object, or NULL if BININDEX is supplied.
See \code{\link[BiocNeighbors]{findKNN}}.}

\item{knn.replace}{logic. Whether to replace knn in FSPY object}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to \code{\link[BiocNeighbors]{findKNN}} function}
}
\value{
An FSPY object with knn, knn.index and knn.distance information.
}
\description{
Calculates and stores a k-nearest neighbor graph based on Euclidean
   distance with (KMKNN) algorithm using log-transformed signaling matrix of
   flow cytometry data. The base function are base on \code{\link[BiocNeighbors]{findKNN}}.
}
\seealso{
\code{\link[BiocNeighbors]{findKNN}}
}
