% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{runKNN}
\alias{runKNN}
\title{Calculate k-nearest neighbors of FSPY}
\usage{
runKNN(object, knn = 30, list.params = list(), knn.replace = T,
  BNPARAM = NULL, verbose = T)
}
\arguments{
\item{object}{an FSPY object}

\item{knn}{numeric, number of k-nearest neighbors}

\item{knn.replace}{logic. Whether to replace knn in FSPY object}

\item{BNPARAM}{A BiocNeighborParam object, or NULL if BININDEX is supplied.
See \code{\link[BiocNeighbors]{findKNN}}.}

\item{verbose}{logic. Whether to print calculation progress.}

\item{iter.max}{numeric. The maximum number of iterations allowed.}

\item{knn.cluster}{logic. Whether to run knn cluster.}
}
\value{
An FSPY object
}
\description{
Calculates and stores a k-nearest neighbor graph based on Euclidean
   distance with (KMKNN) algorithm using log-transformed signaling matrix of
   flow cytometry data. The base function referes to \code{\link[BiocNeighbors]{findKNN}}.
}
\examples{


}
