% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{runPseudotime}
\alias{runPseudotime}
\title{Calculation of Pseudotime}
\usage{
runPseudotime(object, mode = "undirected", ...)
}
\arguments{
\item{object}{An FSPY object}

\item{mode}{character. Specifies how igraph should interpret the supplied matrix.
Possible values are: directed, undirected, upper, lower, max, min, plus.}

\item{...}{Parameters passing to calculation function.}

\item{verbose}{logical. Whether to print calculation progress.}
}
\value{
An FSPY object
}
\description{
calculation of Pseudotime based on KNN
}
