% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{runPseudotime}
\alias{runPseudotime}
\title{Calculation of Pseudotime}
\usage{
runPseudotime(object, mode = "undirected", dim.type = "tsne",
  dim.use = 1:2, verbose = F, ...)
}
\arguments{
\item{object}{An FSPY object}

\item{mode}{character. Specifies how igraph should interpret the supplied matrix.
Possible values are: directed, undirected, upper, lower, max, min, plus.}

\item{dim.type}{character. Type of dimensionality reduction method used to calculate
pseudotime: umap, tsne, dc, pca or raw.}

\item{dim.use}{numeric. Dimensions used to calculate pseudotime}

\item{verbose}{logical. Whether to print calculation progress.}

\item{...}{Parameters passing to calculation function.}
}
\value{
An FSPY object
}
\description{
calculation of Pseudotime based on KNN
}
\examples{

fspy <- runPseudotime(fspy, verbose = T, dim.type = "umap", dim.use = 1:2)
fspy <- runPseudotime(fspy, verbose = T, dim.type = "tsne", dim.use = 1:2)
fspy <- runPseudotime(fspy, verbose = T, dim.type = "dc", dim.use = 1:3)
fspy <- runPseudotime(fspy, verbose = T, dim.type = "pca", dim.use = 1:3)
fspy <- runPseudotime(fspy, verbose = T, dim.type = "raw")


# tSNE plot colored by pseudotime
plot2D(fspy, item.use = c("tSNE_1", "tSNE_2"), category = "numeric",
       size = 1, color.by = "pseudotime") +
       scale_colour_gradientn(colors = c("#F4D31D", "#FF3222","#7A06A0"))
# UMAP plot colored by pseudotime
plot2D(fspy, item.use = c("UMAP_1", "UMAP_2"), category = "numeric",
       size = 1, color.by = "pseudotime") +
       scale_colour_gradientn(colors = c("#F4D31D", "#FF3222","#7A06A0"))


}
