% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap.R
\name{runUMAP}
\alias{runUMAP}
\title{Calculating UMAP}
\usage{
runUMAP(object, umap.config = umap.defaults, verbose = T)
}
\arguments{
\item{object}{an FSPY object}

\item{umap.config}{object of class umap.config. See \code{\link[umap]{umap}}.}

\item{verbose}{logical. Whether to print calculation progress.}
}
\description{
Calculate Uniform Manifold Approximation and Projection in FSPY
}
\examples{

}
